mongodb/requirements.txt
pymongo
python-dotenv
pymongo[srv]


